tauri: onboarding flow updates (#4328)
* Update onboarding screen details

* update login and signup screens
