+++ nym-vpn-app/src/screens/Onboarding/slides/Speed.tsx
@@ -6,7 +6,7 @@ function Speed() {
   return (
     <div className="flex flex-col items-center gap-4">
       <SpeedAsset className="h-full max-h-72 w-fit" />
-      <h1 className="text-2xl text-center text-baltic-sea dark:text-white">
+      <h1 className="text-2xl text-center whitespace-pre-line text-baltic-sea dark:text-white">
         {t('speed.title')}
       </h1>
       <p className="text-center text-base whitespace-pre-line text-iron dark:text-bombay">
