+++ nym-vpn-app/src/screens/login/index.ts
@@ -1,2 +1,1 @@
 export { default as Login } from './Login';
-export { default as PassphraseLogin } from './PassphraseLogin';
