+++ nym-vpn-app/src/screens/sign-up/index.ts
@@ -1,0 +1,1 @@
+export { default as Signup } from './Signup';
